Advanced.Swift.tbt: m/km parsed to ints

+++ AdvancedMap.Swift/Feature Demo/TurnByTurnBanner.swift
@@ -158,16 +158,17 @@ class TurnByTurnBanner: UIView {
         let parsedTime = NSMutableAttributedString()
         
         if (rawDistance > 1000) {
-            parsedDistance.bold(round(rawDistance / 1000).description)
+            parsedDistance.bold(Int(round(rawDistance / 1000)).description)
             // Use different unit of measurement if it's greater one kilometer
             parsedDistance.normal(" km ".uppercased())
             
-            parsedDistance.bold(round(rawDistance.truncatingRemainder(dividingBy: 1000)).description)
+            parsedDistance.bold(Int(round(rawDistance.truncatingRemainder(dividingBy: 1000))).description)
             // Use different unit of measurement if it's greater one kilometer
             parsedDistance.normal(" m".uppercased())
             
         } else {
-            parsedDistance.bold(round(rawDistance).description)
+            
+            parsedDistance.bold(Int(round(rawDistance)).description)
             parsedDistance.normal(" meters".uppercased())
         }
